requirements.txt
django==3.2.7
djangorestframework==3.12.4 
django-environ==0.7.0

python-telegram-bot==13.7
redis==3.5.3
